requirements.txt
google-cloud-language==0.29.0
Flask==0.12.2
gunicorn==19.7.1
